requirements.txt
# Dependencies
python-telegram-bot==20.7
anthropic==0.8.1
pyyaml==6.0.1
aiosqlite==0.19.0
telethon==1.34.0
python-dateutil==2.8.2
coloredlogs==15.0.1
